# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API VisionHub360",
        "description": "API RESTful con JWT y Swagger",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/roles": {
            "get": {
                "tags": [
                    "Roles y Permisos"
                ],
                "summary": "Listar todos los roles con sus permisos",
                "operationId": "af19abbcadaa38a05835689ad61a53ca",
                "responses": {
                    "200": {
                        "description": "Lista de roles con sus permisos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Administrador General"
                                            },
                                            "permissions": {
                                                "type": "array",
                                                "items": {
                                                    "type": "string",
                                                    "example": "ver_usuarios"
                                                }
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/permissions": {
            "get": {
                "tags": [
                    "Roles y Permisos"
                ],
                "summary": "Listar todos los permisos disponibles",
                "operationId": "08a0633c847f2575cd515e9cc5a4816c",
                "responses": {
                    "200": {
                        "description": "Lista de permisos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "ver_clientes"
                                            },
                                            "guard_name": {
                                                "type": "string",
                                                "example": "api"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}